+++ resources/js/components/KinetixFormSchema.vue
@@ -21,6 +21,7 @@ import KinetixWeekPicker from "./KinetixWeekPicker.vue";
 import KinetixDateRangePicker from "./KinetixDateRangePicker.vue";
 import KinetixAddressPicker from "./KinetixAddressPicker.vue";
 import KinetixRichEditor from "./KinetixRichEditor.vue";
+import KinetixNumberField from "./KinetixNumberField.vue";
 import { cn } from "./primitives/cn";
 import { inputClass, textareaClass } from "@/composables/useShadcnVariants";
 
@@ -469,6 +470,16 @@ const moveRepeaterItem = (name: string, index: number, direction: number) => {
           :week-starts-on="comp.weekStartsOn ?? 1"
           :min-value="comp.minValue"
           :max-value="comp.maxValue"
+          @update:value="(v) => emit('update:value', comp.name, v)"
+        />
+
+        <!-- Number field (Reka NumberField with steppers + Intl formatting) -->
+        <KinetixNumberField
+          v-else-if="comp.type === 'number-field'"
+          :value="values[comp.name]"
+          :config="comp.numberConfig"
+          :disabled="comp.isDisabled"
+          :placeholder="comp.placeholder"
           @update:value="(v) => emit('update:value', comp.name, v)"
         />